Fix installation paths.
When installing cutorch locally, the installer tries to copy THC
files outside of the rocks tree (e.g. /usr/lib/libTHC.so or
/usr/include/THC/THC.c). Consequently, a local installation fails
because the user isn't allowed to copy files to /usr.
This patch fixes the installation paths so everything is copied into
rocks trees only, thus fixing the local installation problem.