Remove broken apply, map and map2 functions.
They were copied from torch, using TH_TENSOR_APPLY macros which don't
work with CudaTensors.
If we want to make this work without having to resort to copying data to
the CPU, we'd have to do runtime compilation as suggested in #118. But
for now I think we should just remove these functions here.